✨ Committed Donors Hookup, moving code to common
