+++ components/pages/submission-system/program-dashboard/DASHBOARD_SUMMARY_QUERY.gql
