+++ components/pages/submission-system/program-dashboard/StatsBar/index.tsx
@@ -6,11 +6,15 @@ import Typography from 'uikit/Typography';
 import useTheme from 'uikit/utils/useTheme';
 import Button from 'uikit/Button';
 import Icon from 'uikit/Icon';
-import DASHBOARD_SUMMARY_QUERY from './DASHBOARD_SUMMARY_QUERY.gql';
+import DASHBOARD_SUMMARY_QUERY from '../DASHBOARD_SUMMARY_QUERY.gql';
 import { useQuery } from '@apollo/react-hooks';
 import { usePageQuery } from 'global/hooks/usePageContext';
 import _ from 'lodash';
-import { POLL_INTERVAL_MILLISECONDS } from '../common';
+import {
+  POLL_INTERVAL_MILLISECONDS,
+  DashboardSummaryData,
+  DashboardSummaryDataVariables,
+} from '../common';
 
 const StatDesc = styled('div')`
   display: flex;
@@ -116,35 +120,9 @@ const Statistic: React.ComponentType<{ quantity: String; description: String }> 
   </StatDesc>
 );
 
-type ProgramDonorSummaryStats = {
-  registeredDonorsCount: number;
-  percentageCoreClinical: number;
-  percentageTumourAndNormal: number;
-  donorsProcessingMolecularDataCount: number;
-  filesToQcCount: number;
-  donorsWithReleasedFilesCount: number;
-  allFilesCount: number;
-  fullyReleasedDonorsCount: number;
-  partiallyReleasedDonorsCount: number;
-  noReleaseDonorsCount: number;
-};
-
-type Program = {
-  commitmentDonors: number;
-};
-
-type DasboardSummaryData = {
-  programDonorSummaryStats: ProgramDonorSummaryStats;
-  program: Program;
-};
-
-type DashboardSummaryDataVariables = {
-  programShortName: string;
-};
-
 export default () => {
   const { shortName: programShortName } = usePageQuery<{ shortName: string }>();
-  const { data, loading } = useQuery<DasboardSummaryData, DashboardSummaryDataVariables>(
+  const { data, loading } = useQuery<DashboardSummaryData, DashboardSummaryDataVariables>(
     DASHBOARD_SUMMARY_QUERY,
     {
       variables: { programShortName: programShortName },
