+++ components/pages/submission-system/program-dashboard/common.tsx
@@ -22,4 +22,30 @@ export const DashboardCard: React.ComponentType<{ cardHeight?: string; loading?:
   </Container>
 );
 
+type ProgramDonorSummaryStats = {
+  registeredDonorsCount: number;
+  percentageCoreClinical: number;
+  percentageTumourAndNormal: number;
+  donorsProcessingMolecularDataCount: number;
+  filesToQcCount: number;
+  donorsWithReleasedFilesCount: number;
+  allFilesCount: number;
+  fullyReleasedDonorsCount: number;
+  partiallyReleasedDonorsCount: number;
+  noReleaseDonorsCount: number;
+};
+
+type Program = {
+  commitmentDonors: number;
+};
+
+export type DashboardSummaryData = {
+  programDonorSummaryStats: ProgramDonorSummaryStats;
+  program: Program;
+};
+
+export type DashboardSummaryDataVariables = {
+  programShortName: string;
+};
+
 export const POLL_INTERVAL_MILLISECONDS: number = 3000;
